Some minor corrections to code: use more generic maths functions (eg. atan instead of atanf), and comment out vector at functions, use [] instead. Added extern "C" to the front of prototypes for .c functions in Utils. Added a configure script (with configure.ac) and a Makefile.in to go with it. Some last additions to Guide.tex.
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@107 450e7745-72da-457f-8a74-b07dc0cda934

+++ ATrous/atrous.cc
@@ -7,13 +7,13 @@ int Filter::getNumScales(long length)
 {
   switch(this->filter1D.size()){
   case 5: 
-    return int(logf(float(length-1))/M_LN2) - 1;
+    return int(log(double(length-1))/M_LN2) - 1;
     break;
   case 3:
-    return int(logf(float(length-1))/M_LN2);
+    return int(log(double(length-1))/M_LN2);
     break;
   default:
-    return 1 + int(logf(float(length-1)/float(this->filter1D.size()-1))/M_LN2);
+    return 1 + int(log(double(length-1)/double(this->filter1D.size()-1))/M_LN2);
     break;
   }
 }
